+++ src/main/java/com/program/proyectov1/logic/Poliza.java
@@ -9,7 +9,7 @@ public class Poliza {
     
     private String codigo;
     private Vehiculo vehiculo;
-    private Integer valorSeguro;
+    private Double valorSeguro;
     private String plazoPagos;
     private String fechaInicioVigencia;
     private List<Cobertura> coberturas;
@@ -19,13 +19,13 @@ public class Poliza {
     public Poliza() {
         this.codigo = "";
         this.vehiculo = new Vehiculo();
-        this.valorSeguro = 0;
+        this.valorSeguro = 0.0;
         this.plazoPagos = "";
         this.fechaInicioVigencia = "";
         this.coberturas = new ArrayList<>();
     }
     
-    public Poliza(Vehiculo vehiculo, Integer valorSeguro, String plazoPagos, String fechaInicioVigencia, List<Cobertura> coberturas) {
+    public Poliza(Vehiculo vehiculo, Double valorSeguro, String plazoPagos, String fechaInicioVigencia, List<Cobertura> coberturas) {
         this.codigo = "";
         this.vehiculo = vehiculo;
         this.valorSeguro = valorSeguro;
@@ -34,7 +34,7 @@ public class Poliza {
         this.coberturas = coberturas;
     }
 
-    public Poliza(String codigo, Vehiculo vehiculo, Integer valorSeguro, String plazoPagos, String fechaInicioVigencia, List<Cobertura> coberturas) {
+    public Poliza(String codigo, Vehiculo vehiculo, Double valorSeguro, String plazoPagos, String fechaInicioVigencia, List<Cobertura> coberturas) {
         this.codigo = codigo;
         this.vehiculo = vehiculo;
         this.valorSeguro = valorSeguro;
@@ -59,11 +59,11 @@ public class Poliza {
         this.vehiculo = vehiculo;
     }
 
-    public Integer getValorSeguro() {
+    public Double getValorSeguro() {
         return valorSeguro;
     }
 
-    public void setValorSeguro(Integer valorSeguro) {
+    public void setValorSeguro(Double valorSeguro) {
         this.valorSeguro = valorSeguro;
     }
 
@@ -98,10 +98,10 @@ public class Poliza {
     }
     
     public void precioTotal(){
-        Integer total =0;
+        Double total =0.0;
         for(Cobertura c: this.coberturas){
             if(this.vehiculo.getValor()*(c.getCostoPorcentual()/100.0)>c.getCostoMinimo()){
-                total+=  Double.valueOf(this.vehiculo.getValor()*(c.getCostoPorcentual()/100.0)).intValue();
+                total+= this.vehiculo.getValor()*(c.getCostoPorcentual()/100.0);
             }else{
                 total+= c.getCostoMinimo();
             }
